Added OverviewSimplified diagram to DBWrench schema file
I’ve been long wanting to update our overview diagram, which we lost
from the main branch over a year ago, not only for posters but also as
a teaching aid for those working with YODA files. The intent was to
restructure this diagram so that the tables are ordered (left to right,
top to bottom) according the the YODA file template workflow.